backend/pyproject.toml
[project]
name = "chess-clone-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-chess>=1.999",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "redis[hiredis]>=5.0.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "websockets>=12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
